Characters only have public variables that should be set in Inspector
Ensure only variables that SHOULD be accessible and edible in the Unity Inspector window are marked as "public". Variables marked as public can be modified by any other class/object. To have better control over variables that do not need to be set and modified in the Unity Inspector, set them to private or protected instead and use getter and, if modifiable, setter functions for other classes/objects to get the variable value and set the variable value instead of directly editing this class's variables.