Keep code dry (Re-use same function when possible)
In BoardManager.cs, we perform a lot of very similar checks to see if a coordinate is within a specified range of X and Y values. These checks are often very long, adding unnecessary bulk to our "if" statements and making the code harder to follow.

It is better coding practice to use a single function that performs this check instead of re-adding very similar checks to every if statement with different variable names. Instead, the coordinates and boundaries can be passed in as Vector2 arguments to our new InRange() function that performs the exact same check.

Adding this function and switching our if statements to use this function will keep our code dry and easier to follow instead having duplicate code.